# api/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Backup Manager /api/v1",
    "version": "1.0.0",
    "summary": "The authoritative definition of the versioned /api/v1 boundary.",
    "description": "Issue #166 (B6.2). This document, not the Go handlers and not the TypeScript client, is the definition of /api/v1. Both sides are generated from it (scripts/api/generate.sh) and both are held to it by a drift check (scripts/api/check-contract-drift.sh plus the conformance tests named in docs/api/contract.md). Compatibility and deprecation rules are in docs/api/contract.md."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Every path in this document is relative to the versioned base path. /health/live and /health/ready are deliberately NOT in this contract: they sit outside /api/v1, carry no authentication and exist for an orchestrator, not for a client of this API."
    }
  ],
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "bm_session",
        "description": "The authenticated session. For the generic runtime profile this is apps/common/auth/local's own HTTP-only cookie; a runtime profile with a native gateway (native_auth) establishes the same session from provider identity headers accepted only from a trusted peer. Which mechanism established it is a profile concern; that a caller of a secured operation has one is not."
      },
      "csrf": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "The double-submit CSRF token (apps/common/csrf). Required on every state-changing operation, and on the two read-only-in-effect operations that still open an outbound connection to a caller-supplied host."
      },
      "bootstrapToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Bootstrap-Token",
        "description": "The single-use enrollment secret printed to the runtime's own log. Only /auth/enroll accepts it."
      }
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Caller-supplied key that makes a retried submission a replay rather than a second operation. Reusing one key for two different operation bodies is IDEMPOTENCY_KEY_CONFLICT, never a silent second run.",
        "schema": {
          "type": "string"
        }
      }
    },
    "schemas": {
...
      "SubmitOperationRequest": {
        "type": "object",
        "description": "POST /operations. The idempotency key is a header, not a body field: it is a property of the retry, not of the operation.",
        "additionalProperties": false,
        "properties": {
          "action": {
            "type": "string",
            "description": "What to submit. An enumerated action name, never a command line."
          },
          "config_revision": {
            "type": "string",
            "description": "Optimistic concurrency: the revision the caller believes is current. A stale value is refused with CONFIG_REVISION_STALE and nothing runs."
          }
        },
        "required": [
          "action",
          "config_revision"
        ]
      },
...
      "UpdateSettingsRequest": {
        "type": "object",
        "description": "PATCH /settings. An enumerated request type, never a configuration passthrough.",
        "additionalProperties": false,
        "properties": {
          "retention": {
            "$ref": "#/components/schemas/UpdateRetentionSettings",
            "x-go-pointer": true
          }
        },
        "required": []
      },
...
    },
...
  },
  "paths": {
...
    "/operations": {
      "post": {
        "operationId": "submitOperation",
        "summary": "Submit a durable operation.",
        "tags": [
          "operations"
        ],
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitOperationRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Operation"
                }
              }
            }
          },
          "400": {
            "description": "Refused: INVALID_REQUEST.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "INVALID_REQUEST"
            ]
          },
          "401": {
            "description": "Refused: UNAUTHENTICATED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "UNAUTHENTICATED"
            ]
          },
          "403": {
            "description": "Refused: CSRF_TOKEN_MISSING, CSRF_TOKEN_MISMATCH, DESTRUCTIVE_OPERATIONS_DISABLED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "CSRF_TOKEN_MISSING",
              "CSRF_TOKEN_MISMATCH",
              "DESTRUCTIVE_OPERATIONS_DISABLED"
            ]
          },
          "409": {
            "description": "Refused: CONFIG_REVISION_STALE, IDEMPOTENCY_KEY_CONFLICT, OPERATION_ALREADY_RUNNING. The body is one of two shapes and the code decides which: CONFIG_REVISION_STALE carries the structured config_revision a client retries against, the other two are the plain error envelope. oneOf rather than one schema with an optional config_revision, so a client reading .config_revision after a CONFIG_REVISION_STALE still gets a field the type system guarantees is there.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ConfigRevisionStaleResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    }
                  ]
                }
              }
            },
            "x-error-codes": [
              "CONFIG_REVISION_STALE",
              "IDEMPOTENCY_KEY_CONFLICT",
              "OPERATION_ALREADY_RUNNING"
            ]
          },
          "500": {
            "description": "Refused: INTERNAL.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "INTERNAL"
            ]
          }
        },
        "x-destructive-gate": true,
        "x-idempotency-key": "required",
        "x-optimistic-concurrency": "config_revision",
        "x-csrf-required": true,
        "x-authenticated": true
      },
      "get": {
        "operationId": "listOperations",
        "summary": "Recent operations, newest first. A client with no operation id to poll with has no other way to learn that anything is running.",
        "tags": [
          "operations"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListOperationsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Refused: UNAUTHENTICATED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "UNAUTHENTICATED"
            ]
          },
          "500": {
            "description": "Refused: INTERNAL.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "INTERNAL"
            ]
          }
        },
        "x-destructive-gate": false,
        "x-idempotency-key": "none",
        "x-optimistic-concurrency": null,
        "x-csrf-required": false,
        "x-authenticated": true
      }
    },
...
    "/backup-sets": {
      "get": {
        "operationId": "listBackupSets",
        "summary": "Every configured backup set.",
        "tags": [
          "backup-sets"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListBackupSetsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Refused: UNAUTHENTICATED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "UNAUTHENTICATED"
            ]
          },
          "500": {
            "description": "Refused: INTERNAL.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "INTERNAL"
            ]
          }
        },
        "x-destructive-gate": false,
        "x-idempotency-key": "none",
        "x-optimistic-concurrency": null,
        "x-csrf-required": false,
        "x-authenticated": true
      },
      "post": {
        "operationId": "createBackupSet",
        "summary": "Persist a new backup set, and optionally start its first run.",
        "tags": [
          "backup-sets"
        ],
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBackupSetRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateBackupSetResponse"
                }
              }
            }
          },
          "400": {
            "description": "Refused: INVALID_REQUEST, SSH_KEY_NOT_FOUND.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "INVALID_REQUEST",
              "SSH_KEY_NOT_FOUND"
            ]
          },
          "401": {
            "description": "Refused: UNAUTHENTICATED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "UNAUTHENTICATED"
            ]
          },
          "403": {
            "description": "Refused: CSRF_TOKEN_MISSING, CSRF_TOKEN_MISMATCH, DESTRUCTIVE_OPERATIONS_DISABLED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "CSRF_TOKEN_MISSING",
              "CSRF_TOKEN_MISMATCH",
              "DESTRUCTIVE_OPERATIONS_DISABLED"
            ]
          },
          "500": {
            "description": "Refused: INTERNAL.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "INTERNAL"
            ]
          }
        },
        "x-destructive-gate": false,
        "x-idempotency-key": "none",
        "x-optimistic-concurrency": null,
        "x-csrf-required": true,
        "x-authenticated": true
      }
    },
...
    "/backup-sets/{id}": {
      "get": {
        "operationId": "getBackupSet",
        "summary": "Read one backup set. The id is a two-part source/name identity, so it spans path segments.",
        "tags": [
          "backup-sets"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The two-part identity, e.g. \"nas-a/photos\". Registered in the router as a catch-all for exactly that reason."
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupSet"
                }
              }
            }
          },
          "401": {
            "description": "Refused: UNAUTHENTICATED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "UNAUTHENTICATED"
            ]
          },
          "404": {
            "description": "Refused: BACKUP_SET_NOT_FOUND.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "BACKUP_SET_NOT_FOUND"
            ]
          },
          "500": {
            "description": "Refused: INTERNAL.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "INTERNAL"
            ]
          }
        },
        "x-destructive-gate": false,
        "x-idempotency-key": "none",
        "x-optimistic-concurrency": null,
        "x-csrf-required": false,
        "x-authenticated": true
      }
    },
...
    "/settings": {
      "get": {
        "operationId": "getSettings",
        "summary": "The settings in effect, plus the schema they are validated against.",
        "tags": [
          "settings"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Refused: UNAUTHENTICATED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "UNAUTHENTICATED"
            ]
          },
          "500": {
            "description": "Refused: INTERNAL.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "INTERNAL"
            ]
          }
        },
        "x-destructive-gate": false,
        "x-idempotency-key": "none",
        "x-optimistic-concurrency": null,
        "x-csrf-required": false,
        "x-authenticated": true
      },
      "patch": {
        "operationId": "updateSettings",
        "summary": "Apply only the named fields; everything else keeps what the running configuration says.",
        "tags": [
          "settings"
        ],
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Refused: INVALID_REQUEST.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "INVALID_REQUEST"
            ]
          },
          "401": {
            "description": "Refused: UNAUTHENTICATED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "UNAUTHENTICATED"
            ]
          },
          "403": {
            "description": "Refused: CSRF_TOKEN_MISSING, CSRF_TOKEN_MISMATCH.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "CSRF_TOKEN_MISSING",
              "CSRF_TOKEN_MISMATCH"
            ]
          },
          "500": {
            "description": "Refused: INTERNAL.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "INTERNAL"
            ]
          }
        },
        "x-destructive-gate": false,
        "x-idempotency-key": "none",
        "x-optimistic-concurrency": null,
        "x-csrf-required": true,
        "x-authenticated": true
      }
    },
...
  },
...
}